Add holiday hat overlay for profile image
